Continuing improvements to model.
a) Provided defaults for various constructors and assignment operators
including those of the move specifications.
b) Implemented a swap to make things easier.
c) Added [] and at() functions into key_delegate class as to return the
delegate.
d) Renamed some methods to make it clear that the registering a delegate
instead of a collection of functions.

@todo
i) To be able to register lambdas without requiring to wrap within
std:;function<>.
ii) To be able to register a collection of functions without have to
wrap within a std:::make_tuple.

@future
a) Look into registering functions allow with binding parameters.