Able to get rid of std::function requirement.
Therefore able to register lambda functions directly, but it is required to provide a tuple hint either the index or the element.

Need to find a way to make it possible to register all functions at once in the constructor as well as in the register_functions method.

It might be required to still use std::functions since what occurs for functions that would go out of scope, those need to be copied instead of being a pointer to a function.